Don't stage encrypted AIPs twice, refs #1112
Update `self.current_location` to the temporary staging directory after
copying and decrypting the package.  This is necessary so subsequent
calls to `fetch_local_path()` check if the current (staging) Space is
encrypted (via `is_encrypted()`), not whether the AIPsStore Space
is encrypted.  If `fetch_local_path()` thinks the package is still
encrypted it will copy it to a new temp dir and decrypt it again.

Also:
- Copy encrypted AIPs directly from storage location to temporary
  staging directory and decrypt there
- Add a debug log message to `get_download_path()`